+++ buildTools/linux-x64/mod2gbt/mod2gbt.c
@@ -1,29 +1,9 @@
 /*
- * mod2gbt v2.2 (Part of GBT Player)
+ * mod2gbt v2.5 rulz (Part of GBT Player)
  *
- * Copyright (c) 2009-2018, Antonio Niño Díaz <antonio_nd@outlook.com>
- * All rights reserved.
+ * SPDX-License-Identifier: MIT
  *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * * Redistributions of source code must retain the above copyright notice, this
- *  list of conditions and the following disclaimer.
- *
- * * Redistributions in binary form must reproduce the above copyright notice,
- *   this list of conditions and the following disclaimer in the documentation
- *   and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
- * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
- * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
- * SERVICES * LOSS OF USE, DATA, OR PROFITS * OR BUSINESS INTERRUPTION) HOWEVER
- * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
- * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ * Copyright (c) 2009-2020, Antonio Niño Díaz <antonio_nd@outlook.com>
  */
 
 #include <stdio.h>
@@ -31,13 +11,17 @@
 #include <string.h>
 
 #define DEFAULT_ROM_BANK (2)
+#define BANK_NUM_UNBANKED 0
 
 unsigned int current_output_bank;
 
+int perform_speed_convertion = 1;
 typedef unsigned char u8;
 typedef signed   char s8;
 typedef unsigned short int u16;
 typedef signed   short int s16;
+u8 ch1_inst_preserve = 0;
+u8 ch2_inst_preserve = 0;
 
 #define abs(x) (((x) > 0) ? (x) : -(x))
 #define BIT(n) (1 << (n))
@@ -144,19 +128,13 @@ u8 mod_get_index_from_period(u16 period, int pattern, int step, int channel)
     {
         if (period < mod_period[(6 * 12) - 1])
         {
-            if (channel != 4) // Pitch ignored for noise channel
-            {
-                printf("\nPattern %d, Step %d, Channel %d. Note too high!\n",
-                       pattern, step, channel);
-            }
+            printf("\nPattern %d, Step %d, Channel %d. Note too high!\n",
+                   pattern, step, channel);
         }
         else if (period > mod_period[0])
         {
-            if (channel != 4) // Pitch ignored for noise channel
-            {
-                printf("\nPattern %d, Step %d, Channel %d. Note too low!\n",
-                       pattern, step, channel);
-            }
+            printf("\nPattern %d, Step %d, Channel %d. Note too low!\n",
+                   pattern, step, channel);
         }
     }
     else
@@ -187,6 +165,21 @@ u8 mod_get_index_from_period(u16 period, int pattern, int step, int channel)
 
     return nearest_index;
 }
+
+const u8 gbt_noise[16] = {
+	// 7 bit
+	0x5F,0x4E,0x3E,0x2F,0x2E,0x2C,0x1F,0x0F,
+	// 15 bit
+	0x64,0x54,0x44,0x24,0x00,
+	0x67,0x56,0x46
+};
+/*const u8 gbt_noise[16] = { // Old, Substituted for mathematical matches.
+	// 7 bit
+	0x5F,0x5B,0x4B,0x2F,0x3B,0x58,0x1F,0x0F,
+	// 15 bit
+	0x90,0x80,0x70,0x50,0x00,
+	0x67,0x63,0x53
+};*/
 
 //------------------------------------------------------------------------------
 //--                                                                          --
@@ -282,8 +275,10 @@ int volume_mod_to_gb_ch3(int v) // Channel 3
 
 int speed_mod_to_gb(int s)
 {
-    // Amiga's 50 Hz to GB's 60 Hz
-    return (s * 60) / 50;
+    if (perform_speed_convertion) // Amiga's 50 Hz to GB's 60 Hz
+        return (s * 60) / 50;
+    else
+        return s;
 }
 
 // Returns 1 if ok
@@ -295,8 +290,39 @@ int effect_mod_to_gb(u8 pattern_number, u8 step_number, u8 channel,
     {
         case 0x0: // Arpeggio
         {
-            *converted_num = 1;
+            if (effectparams != 0) // Arp has any data
+            {
+                *converted_num = 1;
+                *converted_params = effectparams;
+                return 1;
+            } else  {               // Mistook no effets for arp 000,
+                *converted_num = 7;  // use No Op NOP trigger effect instead.
+                *converted_params = effectparams;
+                return 1;
+            }
+        }
+        case 0x1:   // Ch1,2,3 Pitch Slide UP
+        {
+            *converted_num = 4;
+            *converted_params = (effectparams & 0x7F);
+            return 1;
+        }
+        case 0x2:   // Ch1,2,3 Pitch Slide DOWN
+        {
+            *converted_num = 4;
+            *converted_params = ( (effectparams & 0x7F) | 0x80 );
+            return 1;
+        }
+        case 0x9:   // Ch1,2,4 Volume + envelope, direct NRx2, was Offset
+        {
+            *converted_num = 15;
             *converted_params = effectparams;
+            return 1;
+        }
+        case 0xA:   // Volume Slide (Volume envelope)
+        {
+            *converted_num = 7; // NOP
+            *converted_params = 1;
             return 1;
         }
         case 0xB: // Jump
@@ -399,6 +425,17 @@ void convert_channel1(u8 pattern_number, u8 step_number, u8 note_index,
     int command_len = 1; // NOP
 
     u8 instrument = samplenum & 3;
+    if (samplenum == 0) {
+        instrument = ch1_inst_preserve;
+        // printf("\nReplaced ch1 Instument %d with %d", samplenum, instrument);
+    }
+    ch1_inst_preserve = instrument;
+    if (samplenum > 4)
+        {
+            printf("\nWarning: Channel 1 must use Pulse waves 1-4, "
+                    "but found Instument %d, at Pattern %d, step %d.",
+                    samplenum, pattern_number, step_number);
+        }
 
     if (note_index > (6 * 12 - 1)) // Not valid note -> check if any effect
     {
@@ -417,7 +454,7 @@ void convert_channel1(u8 pattern_number, u8 step_number, u8 note_index,
                 u8 converted_num, converted_params;
                 if (effect_mod_to_gb(pattern_number, step_number, 1, effectnum,
                                      effectparams, &converted_num,
-                                     &converted_params) == 1)
+                                     &converted_params) == 1 && effectnum != 0xA)
                 {
                     result[0] = BIT(6) | (instrument << 4) | converted_num;
                     result[1] = converted_params;
@@ -425,7 +462,7 @@ void convert_channel1(u8 pattern_number, u8 step_number, u8 note_index,
                 }
                 else
                 {
-                    if (effectnum != 0)
+                    if (effectnum != 0 && effectnum != 0xA)
                     {
                         printf("Invalid command at pattern %d, step %d, channel"
                                " 1: %01X%02X\n", pattern_number, step_number,
@@ -461,9 +498,10 @@ void convert_channel1(u8 pattern_number, u8 step_number, u8 note_index,
                                  effectparams, &converted_num,
                                  &converted_params) == 1)
             {
-                // Note + Effect
+                // Note + Effect + Is instrument?
                 result[0] = BIT(7) | note_index;
-                result[1] = BIT(7) | (instrument << 4) | converted_num;
+                result[1] = BIT(7) |
+                (samplenum = 0 ? BIT(6) : (instrument << 4)) | converted_num;
                 result[2] = converted_params;
                 command_len = 3;
             }
@@ -502,6 +540,18 @@ void convert_channel2(u8 pattern_number, u8 step_number, u8 note_index,
     int command_len = 1; // NOP
 
     u8 instrument = samplenum & 3;
+    if (samplenum == 0) {
+        instrument = ch2_inst_preserve;
+        // printf("\nReplaced ch2 Instument %d with %d", samplenum, instrument);
+    }
+    ch2_inst_preserve = instrument;
+    if (samplenum > 4)
+        {
+            printf("\nWarning: Channel 2 must use Pulse waves 1-4, "
+                    "but found Instument %d, at Pattern %d, step %d.",
+                    samplenum, pattern_number, step_number);
+
+        }
 
     if (note_index > (6 * 12 - 1)) // Not valid note -> check if any effect
     {
@@ -520,7 +570,7 @@ void convert_channel2(u8 pattern_number, u8 step_number, u8 note_index,
                 u8 converted_num, converted_params;
                 if (effect_mod_to_gb(pattern_number, step_number, 2, effectnum,
                                      effectparams, &converted_num,
-                                     &converted_params) == 1)
+                                     &converted_params) == 1 && effectnum != 0xA)
                 {
                     result[0] = BIT(6) | (instrument << 4) | converted_num;
                     result[1] = converted_params;
@@ -528,7 +578,7 @@ void convert_channel2(u8 pattern_number, u8 step_number, u8 note_index,
                 }
                 else
                 {
-                    if (effectnum != 0)
+                    if (effectnum != 0 && effectnum != 0xA)
                     {
                         printf("Invalid command at pattern %d, step %d, channel"
                                " 2: %01X%02X\n", pattern_number, step_number,
@@ -564,9 +614,10 @@ void convert_channel2(u8 pattern_number, u8 step_number, u8 note_index,
                                  effectparams, &converted_num,
                                  &converted_params) == 1)
             {
-                // Note + Effect
+                // Note + Effect + Is instrument?
                 result[0] = BIT(7) | note_index;
-                result[1] = BIT(7) | (instrument << 4) | converted_num;
+                result[1] = BIT(7) |
+                (samplenum = 0 ? BIT(6) : (instrument << 4)) | converted_num;
                 result[2] = converted_params;
                 command_len = 3;
             }
@@ -612,7 +663,7 @@ void convert_channel3(u8 pattern_number, u8 step_number, u8 note_index,
             if (effectnum == 0xC)
             {
                 // Volume
-                result[0] = BIT(5) | volume_mod_to_gb_ch3(effectparams);
+                result[0] = BIT(5) | (volume_mod_to_gb_ch3(effectparams) << 1);
                 command_len = 1;
             }
             else
@@ -621,7 +672,8 @@ void convert_channel3(u8 pattern_number, u8 step_number, u8 note_index,
                 u8 converted_num, converted_params;
                 if (effect_mod_to_gb(pattern_number, step_number, 3, effectnum,
                                      effectparams, &converted_num,
-                                     &converted_params) == 1)
+                                     &converted_params) == 1 &&
+                                     effectnum != 0xA && effectnum != 0x9)
                 {
                     result[0] = BIT(6) | converted_num;
                     result[1] = converted_params;
@@ -652,6 +704,16 @@ void convert_channel3(u8 pattern_number, u8 step_number, u8 note_index,
     else // New note
     {
         u8 instrument = (samplenum - 8) & 15; // Only 0-7 implemented
+        if (samplenum < 8)          // 1-7 is now 8-15, tb implemented
+        {
+            instrument = (samplenum + 7) & 15;
+        }
+        if (samplenum > 16)
+        {
+            printf("\nWarning: Channel 3 must use Waves 8-15, "
+                    "but found Instument %d, at Pattern %d, step %d.",
+                    samplenum, pattern_number, step_number);
+        }
 
         u8 converted_num, converted_params;
         if (effectnum == 0xC)
@@ -734,7 +796,7 @@ void convert_channel4(u8 pattern_number, u8 step_number, u8 note_index,
                 u8 converted_num, converted_params;
                 if (effect_mod_to_gb(pattern_number, step_number, 4, effectnum,
                                      effectparams, &converted_num,
-                                     &converted_params) == 1)
+                                     &converted_params) == 1 && effectnum != 0xA)
                 {
                     result[0] = BIT(6) | converted_num;
                     result[1] = converted_params;
@@ -742,7 +804,7 @@ void convert_channel4(u8 pattern_number, u8 step_number, u8 note_index,
                 }
                 else
                 {
-                    if (effectnum != 0)
+                    if (effectnum != 0 && effectnum != 0xA)
                     {
                         printf("Invalid command at pattern %d, step %d, channel"
                                " 4: %01X%02X\n", pattern_number, step_number,
@@ -762,16 +824,37 @@ void convert_channel4(u8 pattern_number, u8 step_number, u8 note_index,
             command_len = 1;
         }
     }
-    else // New note (not a real note...)
+    else // New note (noise) NR43 SSSS WDDD Clock Shift, Width mode of LFSR, Divisor code
     {
-        u8 instrument = (samplenum - 16) & 0x1F; // Only 0 - 0xF implemented
+        u8 instrument = gbt_noise[((samplenum - 16) & 0x1F)]; // Only 0 - 0xF implemented
+        u8 noise_break = 0;
+        u8 noise = 0;
+        if (samplenum < 16)
+        {
+            printf("\nWarning: Channel 4 must use Noises 16-31, "
+                    "but found Instument %d at Pattern %d, step %d.",
+                    samplenum, pattern_number, step_number);
+        }
+        // This makes a smooth Ramp of every noise type, inspired by Pigu-A's Cherry Blossom Dive
+        // SSSS WDDD, preserve Width bit, combine Shift + Divisor (ignore bit 0100), add pitch.
+        // Divisor 4,5,6,7 can make any noise found in 0,1,2,3 unless with 0 Clock Shift.
+        // Solution, add 4, add note, if less than 4, set bit 0x04 to 0, and remove 4 again.
+        // Notes will pitch correctly using C D# F# A# C, scale has been divided by 3.
+        if (samplenum < 32 && samplenum > 16) // Noise
+        {
+            noise_break = ( (instrument & 0x03) | (((instrument & 0xF0) >> 2) + 4) );
+            noise_break = noise_break - (((note_index + 1) / 3) - 8);
+            noise = ( (noise_break & 0x03) |
+            ((((noise_break - 4) < 0 ? 0x0 : (noise_break - 4)) & 0x3C) << 2) |
+             (noise_break > 3 ? 0x04 : 0x0) ) | (instrument & 0x08);
+        }
 
         u8 converted_num, converted_params;
         if (effectnum == 0xC)
         {
             // Note + Volume
-            result[0] = BIT(7) | instrument;
-            result[1] = volume_mod_to_gb(effectparams);
+            result[0] = BIT(7) | (0x7F & noise);
+            result[1] = ( (noise & BIT(7)) << 1 ) | volume_mod_to_gb(effectparams);
             command_len = 2;
         }
         else
@@ -781,8 +864,8 @@ void convert_channel4(u8 pattern_number, u8 step_number, u8 note_index,
                                  &converted_params) == 1)
             {
                 // Note + Effect
-                result[0] = BIT(7) | instrument;
-                result[1] = BIT(7) | converted_num;
+                result[0] = BIT(7) | (0x7F & noise);
+                result[1] = BIT(7) | ( (noise & BIT(7)) << 1 ) | converted_num;
                 result[2] = converted_params;
                 command_len = 3;
             }
@@ -882,8 +965,9 @@ void convert_pattern(_pattern_t *pattern, u8 number)
 
 void print_usage(void)
 {
-    printf("Usage: mod2gbt modfile.mod label_name [N]\n");
-    printf("       N: Set output to ROM bank N (defaults to %d).",
+    printf("Usage: mod2gbt modfile.mod label_name [N] [-speed]\n");
+    printf("       -speed      Don't convert speed from 50 Hz to 60 Hz.\n");
+    printf("       N: Set output to ROM bank N (defaults to %d, use 0 for unbanked)",
            DEFAULT_ROM_BANK);
     printf("\n\n");
 }
@@ -892,23 +976,23 @@ int main(int argc, char *argv[])
 {
     int i;
 
-    printf("mod2gbt v2.2 (part of GBT Player)\n");
-    printf("Copyright (c) 2009-2018 Antonio Niño Díaz "
+    printf("mod2gbt v2.5 rulz (part of GBT Player)\n");
+    printf("Copyright (c) 2009-2020 Antonio Niño Díaz "
            "<antonio_nd@outlook.com>\n");
     printf("All rights reserved\n");
     printf("\n");
 
-    if ((argc < 3) || (argc > 4))
+    if ((argc < 3) || (argc > 6))
     {
         print_usage();
         return -1;
     }
 
-    strcpy(label_name, argv[2]);
+    strncpy(label_name, argv[2], sizeof(label_name));
 
     current_output_bank = DEFAULT_ROM_BANK;
 
-    if (argc == 4)
+    if (argc == 4 || argc == 5)
     {
         if (sscanf(argv[3], "%d", &current_output_bank) != 1)
         {
@@ -918,7 +1002,20 @@ int main(int argc, char *argv[])
         }
         else
         {
-            printf("Output to bank: %d\n", current_output_bank);
+            if (current_output_bank == BANK_NUM_UNBANKED) {
+                printf("Bank set to 0, so output will be unbanked\n");
+            } else {
+                printf("Output to bank: %d\n", current_output_bank);
+            }
+        }
+    }
+
+    for (i = 4; i < argc; i++)
+    {
+        if (strcmp(argv[i], "-speed") == 0)
+        {
+            perform_speed_convertion = 0;
+            printf("Disabled speed convertion.\n\n");
         }
     }
 
@@ -960,9 +1057,11 @@ int main(int argc, char *argv[])
 
     out_write_str("\n// File created by mod2gbt\n\n");
 
-    out_write_str("#pragma bank=");
-    out_write_dec(current_output_bank);
-    out_write_str("\n\n");
+    if (current_output_bank != BANK_NUM_UNBANKED) {
+        out_write_str("#pragma bank=");
+        out_write_dec(current_output_bank);
+        out_write_str("\n\n");
+    }
 
     printf("\nConverting patterns...\n");
     for (i = 0; i < num_patterns; i++)
@@ -973,7 +1072,7 @@ int main(int argc, char *argv[])
 
     printf("\n\nPattern order...\n");
 
-    out_write_str("const unsigned char * const");
+    out_write_str("const unsigned char * const ");
     out_write_str(label_name);
     out_write_str("_Data[] = {\n");